deep_galerkin_project/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dgmlib"
version = "0.1.0"
description = "Production-grade Deep Galerkin Model for solving PDEs in quantitative finance"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Quant ML Engineering Team"}
]
keywords = ["deep-learning", "pde", "quantitative-finance", "option-pricing", "physics-informed-neural-networks"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "tensorboard>=2.13.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "streamlit>=1.25.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.82.0",
    "ruff>=0.0.280",
    "black>=23.7.0",
    "mypy>=1.4.0",
    "isort>=5.12.0",
]
rl = [
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.1.0",
]

[project.scripts]
dgm-train = "scripts.train:main"
dgm-price = "scripts.price_cli:main"

[tool.setuptools.packages.find]
include = ["dgmlib*", "scripts*", "api*", "ui*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
exclude = [".git", ".venv", "build", "dist"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=dgmlib --cov-report=term-missing --cov-report=html"
